Safari-611.1.10/LayoutTests/accessibility/mac/malformed-table-expected.txt

Table with CSS and ARIA
Heading one	Heading two
a	b
Table with CSS and no ARIA
Heading one	Heading two
a	b
This tests that contents in malformed tables are accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table1.role is 'AXRole: AXTable'
PASS table2.role is 'AXRole: AXTable'
PASS table1ColumnCount is 2
PASS table1RowCount is 2
PASS table2ColumnCount is 2
PASS table2RowCount is 2

Check rows for Table1
PASS rowa.isEqual(rowb) is true
PASS rowa.role is 'AXRole: AXRow'
PASS rowa.isEqual(rowb) is true
PASS rowa.role is 'AXRole: AXRow'

Check rows for Table2
PASS rowa.isEqual(rowb) is true
PASS rowa.role is 'AXRole: AXRow'
PASS rowa.isEqual(rowb) is true
PASS rowa.role is 'AXRole: AXRow'

Check cells for Table1
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'

Check cells for Table2
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS cella.isEqual(cellb) is true
PASS cella.role is 'AXRole: AXCell'
PASS successfullyParsed is true

TEST COMPLETE
